Create README.md
To-Do List App is a kind of app that generally used to maintain our day-to-day tasks or list everything that we have to do, with the most important tasks at the top of the list, and the least important tasks at the bottom. The To-do List project is a web application built with HTML,
CSS, and JavaScript.
It records your daily task and to-do lists. You will add a task
for a specific time and also edit it and even mark it as
complete as per your need. Also, you can simply delete any
one of the tasks from the list.